requirements.txt
streamlit
gspread
pandas
google-generativeai
Pillow
